requirements.txt
launchpadlib==2.1.0
fastapi==0.116.1
jira==3.10.5
uvicorn==0.30.6
PyYAML==6.0.2
